added sitting rooms and kitchens
