+++ src/dungeon/room_decorators/mod.rs
@@ -1,0 +1,13 @@
+mod decorate_bedroom;
+mod decorate_kitchen;
+mod decorate_mess_hall;
+pub mod decorate_room;
+mod decorate_sitting_room;
+pub mod room_part;
+pub mod room_type;
+mod utils;
+mod common;
+
+pub use room_part::RoomPart;
+pub use room_type::RoomType;
+pub use decorate_room::decorate_room;
